+++ src/constants.rb
@@ -1,0 +1,7 @@
+module Constants
+  DOMAIN = <YOUR-SHOP-DOMAIN> # e.g <shop>.myshopify.com
+  ACCESS_TOKEN =  <YOUR-ACCESS-TOKEN> #Put your access token here. You can see the docs to see how can the info here https://github.com/Shopify/shopify_api#3-requesting-access-from-a-shop
+  API_VERSION_USING_PAGE = '2019-04'
+  API_VERSION_USING_CURSOR_RELATIVE = '2019-10'
+  LIMIT_PER_PAGE = 100
+end
